refactor: removed existing tests from java_programs to java_testcases
