+++ java_testcases/DETECT_CYCLE_TEST.java
@@ -1,5 +1,8 @@
-package java_programs;
+package java_testcases;
 import java.util.*;
+
+import java_programs.DETECT_CYCLE;
+import java_programs.Node;
 
 
 /**
